Create README.md
The Credit Management System is a Website based application that takes care of the complete internal Card Processing activities of a Bank (management and maintenance of cards).  It is suitable for the management of Debit Cards, Credit Cards as well as Smart Cards.  It mainly involves the maintenance and management of card – related information in a Bank. This product caters to all the Credit, Debit and Smart Card based functions of a bank like Registration of Customers for card issue, Opening Card Accounts for the registered customer, Creation of Cards for registered customers, Hot Carding, making the Card ready for Embossing by creation of Card Data file etc. The Card Management System could interact with an external Card Transaction System, thereby making the Card-related information maintained by the bank, up-to-date.  The purpose is to build a Card Management System which provides complete card processing, to meet the needs of full-fledged Credit card and Debit card based ATM or Point of Sale network.